+++ UIReact/UICore/_executor/_Markup/Vdom/Refs/CreateEventRef.ts
@@ -1,13 +1,13 @@
 import { IWasabyEvent } from 'UICommon/Events';
-import { setEventHook } from 'UICore/Events';
-import { Responsibility, IResponsibilityHandler, } from 'UICore/Ref';
+import { WasabyEvents } from 'UICore/Events';
+import { Responsibility, IResponsibilityHandler } from 'UICore/Ref';
 
 export class CreateEventRef extends Responsibility {
     private tagName: string;
     private eventsObject: {
         events: Record<string, IWasabyEvent[]>;
     };
-    constructor(tagName, eventsObject) {
+    constructor(tagName: string, eventsObject: {events: Record<string, IWasabyEvent[]>;}) {
         super();
         this.tagName = tagName;
         this.eventsObject = eventsObject;
@@ -16,7 +16,7 @@ export class CreateEventRef extends Responsibility {
     getHandler(): IResponsibilityHandler {
         return (node: HTMLElement): void => {
             if (node && Object.keys(this.eventsObject.events).length > 0) {
-                setEventHook(this.tagName, this.eventsObject, node);
+                WasabyEvents.getInstance().setEventHook(this.tagName, this.eventsObject, node);
             }
         };
     }
